GANTT Chart v1.0
Created this and just took some estimates for how long each task will
take.  We should try to stick to the chart as closely as possible so
that we can stay on track. It gets pretty close to submission date.

Also it is a .mpp file so you need project to view, i also put in a PDF
file but it comes out pretty badly, i recommend project.

This is a first version so remember changes can be made, if you guys
read through it and can think of anything that should be added/removed
etc just tell me.

Also moved meeting minutes to a different folder.